Move account login to bg
