+++ src/constants.ts
@@ -25,6 +25,7 @@ export enum MESSAGE_TYPE {
   SAVE_TO_FILE,
   IMPORT_MNEMONIC,
   IMPORT_MNEMONIC_FAILURE,
+  ACCOUNT_LOGIN,
 }
 
 export enum STORAGE {
