+++ src/stores/WalletStore.ts
@@ -1,5 +1,5 @@
 import { Wallet, Insight } from 'qtumjs-wallet';
-import { observable, action, computed, runInAction } from 'mobx';
+import { observable, action, computed } from 'mobx';
 import { find, isEmpty } from 'lodash';
 
 import AppStore from './AppStore';
@@ -19,8 +19,6 @@ const INIT_VALUES = {
 };
 
 export default class WalletStore {
-  private static SCRYPT_PARAMS_PRIV_KEY: any = { N: 8192, r: 8, p: 1 };
-
   @observable public loading = INIT_VALUES.loading;
   @observable public appSalt?: Uint8Array = INIT_VALUES.appSalt;
   @observable public passwordHash?: string = INIT_VALUES.passwordHash;
@@ -44,49 +42,10 @@ export default class WalletStore {
   }
   public wallet?: Wallet = INIT_VALUES.wallet;
 
-  @computed private get validPasswordHash(): string {
-    if (!this.passwordHash) {
-      throw Error('passwordHash should be defined');
-    }
-    return this.passwordHash!;
-  }
   private app: AppStore;
 
   constructor(app: AppStore) {
     this.app = app;
-  }
-
-  /*
-  * Finds the account based on the name and logs in.
-  * @param accountName {string} The account name to search by.
-  */
-  @action
-  public async loginAccount(accountName: string) {
-    this.loading = true;
-
-    runInAction(async () => {
-      const accounts = this.app.networkStore.isMainNet ? this.mainnetAccounts : this.testnetAccounts;
-      const foundAccount = find(accounts, { name: accountName });
-
-      if (!foundAccount) {
-        throw Error('Account should not be undefined');
-      }
-
-      this.loggedInAccount = foundAccount;
-
-      // Recover wallet
-      const network = this.app.networkStore.network;
-      this.wallet = await network.fromEncryptedPrivateKey(
-        this.loggedInAccount!.privateKeyHash,
-        this.validPasswordHash,
-        WalletStore.SCRYPT_PARAMS_PRIV_KEY,
-      );
-
-      // Save logged in account info to local storage for QryptoRpcProvider usage
-      this.setLoggedInAccountToStorage();
-
-      await this.onAccountLoggedIn();
-    });
   }
 
   @action
@@ -117,20 +76,6 @@ export default class WalletStore {
       // Accounts found, route to Account Login page
       chrome.runtime.sendMessage({ type: MESSAGE_TYPE.LOGIN_SUCCESS_WITH_ACCOUNTS });
     }
-  }
-
-  private setLoggedInAccountToStorage() {
-    return chrome.storage.local.set(
-      {
-        [STORAGE.LOGGED_IN_ACCOUNT]: {
-          isMainNet: this.app.networkStore.isMainNet,
-          name: this.loggedInAccount!.name,
-          privateKeyHash: this.loggedInAccount!.privateKeyHash,
-          passwordHash: this.passwordHash,
-        },
-      },
-      () => console.log('loggedInAccount saved in storage'),
-    );
   }
 
   private removeLoggedInAccountFromStorage() {
